#92 [REFACTOR] 전체 채팅 페이지: 훅 분리, 전반적인 코드 리팩
