+++ src/pages/GlobalChatPage/hooks/useGetGlobalChatLogs.ts
@@ -1,7 +1,8 @@
 import { fetchGlobalChatLogs } from "@/pages/GlobalChatPage/apis/globalChat";
 import { useQuery } from "@tanstack/react-query";
 
-export const useGlobalChatLogs = (serverId: number | null) => {
+// 채팅 로그 조회 API 훅
+export const useGetGlobalChatLogs = (serverId: number | null) => {
   return useQuery({
     queryKey: ["dmLogs", serverId],
     queryFn: () => (serverId !== null ? fetchGlobalChatLogs(serverId) : Promise.reject("serverId is null")),
